Till-Zainab/ASE/src/requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0

# Langchain/langgraph
langchain>=0.1.0
langgraph>=0.0.11
langchain-core>=0.1.0
langchain-openai>=0.1.0
openai>=1.3.0

# Coverage analysis
coverage>=6.6.0

# Utils
typing-extensions>=4.5.0
